Fix: Ensurer timer is stopped and not replaced (#37)
# Description

This PR ensures there's no race condition when using the tumbling window
function. Instead of setting the timer to nil when done we just ensure
to stop it.

# Related Issues
- Closes #32 


# Changes summary

<!--
Not mandatory, but helpful if PR is big enough.
Summary for the changes done in this PR. Using bullet list is a good
idea.
* [CHORE] change1
* [FEAT] change1
* [FIX] change1
-->

# QA/testing plan

(Provide information on how to test your changes.)

- [ ] Skip QA (if checked, provide a reason below).
- [ ] No code change (if checked, provide a reason below).
- [ ] Automated integration testing.

# Reviewer's Checklist

- [ ] Ran linter (_make lint_).
- [ ] Ran security scan (_make scan\:sca_).
- [ ] All automated tests passed.
- [ ] Updated godoc (_make docs_).
- [ ] Updated examples.